+++ src/main/java/servicios/CategoriaDAO.java
@@ -4,7 +4,7 @@ import tropa.Categoria;
 
 public class CategoriaDAO extends DAO{
 
-	public Categoria obtenerCategoria(int id){
+	public Categoria obtenerCategoria(Long id){
 		return (Categoria)this.obtener(Categoria.class, id);
 	}
